Updated the mode to work only on s-expressions
It now handles (un)commenting of #_ in front of s-expressions
or (comment) type comment around a region, all with a single function.

Handle (comment...) & #_ comments on regions

Decide what comment to use on region, based on what kind of region is
selected. If it's s-expression, then #_ at the start of the
s-expression will suffice. Otherwise, use (comment ...).